Added `volatile` volume option
Implement volatile volumes. A volatile volume loses the changes made in it as soon as it is unmounted from the last container that was using it. It continues to exist but on the next mount it looks like a fresh docker-on-top volume, that is, it's initial contents is identical to the base dir's contents.

Fixes #3