Implemented the activemounts/ logic
The plugin now determines the active mounts for a volume (not overlay) using the activemounts/ subdirectory corresponding to the volume. It resolves the race condition problem with parallel mounts/unmounts on the same volume.

Gets us closer to fixing #15! However, not there yet, as we are still using go's `map`, which is thread-unsafe.