Attempt to unmount volume on remove to recover from previous failure
This enables more careful treatment of volume removal:
we begin by making sure that a volume's mountpoint no longer exists
or, if it does, we unmount it (force+detach).

If that succeeds, we can then safely remove the volume (which would
fail otherwise, leaving the volume in an inconsistent state, neither
allowing to mount it nor remove it).

If unmounting the mountpoint fails, we refuse to do anything else to
the volume, leaving it exactly in the state it was before.

Update relevant parts of code accordingly. Require in tests that
volumes are removed even when an overlay remains mounted.